requirements.txt
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
flake8==4.0.1
fonttools==4.28.5
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
isort==5.10.1
jedi==0.17.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
numpy==1.22.0
packaging==21.3
parso==0.7.1
Pillow==9.0.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.12.2
pymongo==4.1.0
pyparsing==3.0.6
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
requests==2.27.1
rope==0.23.0
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.9
wrapt==1.13.3
yapf==0.32.0
